updated overall class structure & class diagram

+++ src/main/java/org/example/Requirement.java
@@ -11,9 +11,9 @@ public class Requirement {
     @JsonAlias({"CardType"})
     private CardType cardType;
     @JsonAlias({"MinimumDamage"})
-    private Integer minDamage;
+    private Double minDamage;
 
-    public Requirement(CardType cardType, Integer minDamage) {
+    public Requirement(CardType cardType, Double minDamage) {
         this.cardType = cardType;
         this.minDamage = minDamage;
     }
